chore(refactor): further refactor of upload method

+++ app/controllers/dashboard_controller.rb
@@ -1,3 +1,8 @@
+require 'fileutils'
+require 'json'
+require 'zip_file_generator'
+require 'open-uri'
+
 class DashboardController < ApplicationController
     def index
         @items = Item.all
@@ -5,73 +10,68 @@ class DashboardController < ApplicationController
     end
 
     def upload
-        require 'fileutils'
-        require 'json'
-        require 'zip_file_generator'
-
         uploaded_file = params[:file]
         texture_name = params[:name]
         selection = params[:selection]
-
+        item = selection
         selection = selection + ".json"
-
         collection = "custom"
         examples = "public/resources/example_models/"
         models = "storage/#{collection}/assets/minecraft/models/item/"
         model_out = "storage/#{collection}/assets/minecraft/models/item/#{collection}/"
         img_out = "storage/#{collection}/assets/minecraft/textures/item/#{collection}/"
 
-        unless Dir.exists?(model_out)
-            FileUtils.mkdir_p(model_out)
-        end
-
-        unless Dir.exists?(img_out)
-            FileUtils.mkdir_p(img_out)
-        end
-
-        unless File.exist?(models + selection)
-            FileUtils.cp(examples + selection, models)
-        end
-
+        # Create correct folder structure
+        FileUtils.mkdir_p(model_out) unless Dir.exists?(model_out)
+        FileUtils.mkdir_p(img_out) unless Dir.exists?(img_out)
+        FileUtils.cp(examples + selection, models) unless File.exist?(models + selection)
         unless File.exist?(model_out + texture_name + ".json")
             FileUtils.cp(examples + selection, model_out)
             File.rename(model_out + selection, model_out + texture_name + ".json")
         end
 
+        # Determine core data
+        model_json = JSON.parse(File.read(models + selection))
+        model_no = next_model_number(model_json)
+
+        # Saves updated base model to folder structure
+        output = new_override(model_json, collection, texture_name, model_no)
+        File.write(models + selection, JSON.pretty_generate(output))
+
+        # Saves new child (override) model to folder structure
+        model_out_json = JSON.parse(File.read(model_out + texture_name + ".json"))
+        texture_data = {"textures"=>{"layer0"=>"minecraft:item/#{collection}/#{texture_name}"}}
+        output = model_out_json.merge(texture_data)
+        File.write(model_out + texture_name + ".json", JSON.pretty_generate(output))
+
+        # Saves their image into the folder structure
         File.open(img_out + texture_name + ".png", 'wb') do |file|
             file.write(uploaded_file.read)
         end
 
-        model_json = JSON.parse(File.read(models + selection))
-        
+        # Zips Collections
+        zip_collection(collection)
+
+        # Redirects to Output
+        redirect_to output_dashboard_index_path(uid: model_no, name: texture_name, item: item)
+    end
+
+    def new_override(model_json, collection, texture_name, model_no)
         if has_overrides?(model_json)
-            model_no = next_model_number(model_json)
-            overrides = overrides | model_data(model_no, collection, texture_name)
-            overrides = {"overrides" => overrides}
-            output = model_json.merge(overrides)
+            new_override = model_json["overrides"] | model_data(model_no, collection, texture_name)
+            overrides = {"overrides" => new_override}
+            model_json.merge(overrides)
         else
-            model_no = 1
-            output = model_json.merge({"overrides" => model_data(model_no, collection, texture_name)})
+            model_json.merge({"overrides" => model_data(next_model_number(model_json), collection, texture_name)})
         end
+    end
 
-        File.write(models + selection, JSON.pretty_generate(output))
-        model_json = JSON.parse(File.read(model_out + texture_name + ".json"))
-        texture_data = {"textures"=>{"layer0"=>"minecraft:item/#{collection}/#{texture_name}"}}
-        output = model_json.merge(texture_data)
-
-        File.write(model_out + texture_name + ".json", JSON.pretty_generate(output))
-
+    def zip_collection(collection)
         folder = "storage/#{collection}/"
         zipfile = "public/packs/custom-rp.zip"
-
-        if File.exists?(zipfile)
-            File.delete(zipfile)
-        end
-
+        File.delete(zipfile) if File.exists?(zipfile)
         zf = ZipFileGenerator.new(folder, zipfile)
         zf.write
-
-        redirect_to output_dashboard_index_path(uid: model_no, name: texture_name, item: selection)
     end
 
     def model_data(model_no, collection, texture_name)
@@ -83,17 +83,14 @@ class DashboardController < ApplicationController
     end
 
     def next_model_number(model_json)
-        overrides = model_json["overrides"]
-        predicate = overrides.last
+        return 1 unless has_overrides?(model_json)
+        predicate = model_json["overrides"].last
         model = predicate["predicate"]
-        model["custom_model_data"].to_int + 1 
+        model["custom_model_data"].to_int + 1
     end
 
     def download
-        require 'open-uri'
-        
         filepath = "public/packs/custom-rp.zip"
-
         send_file filepath
     end
